+++ component/libfuse/libfuse2_handler.go
@@ -106,6 +106,7 @@ func (lf *Libfuse) convertConfig() *C.fuse_options_t {
 	fuse_opts.readonly = C.bool(lf.readOnly)
 	fuse_opts.allow_other = C.bool(lf.allowOther)
 	fuse_opts.trace_enable = C.bool(lf.traceEnable)
+	fuse_opts.non_empty = C.bool(lf.nonEmptyMount)
 	return fuse_opts
 }
 
@@ -197,6 +198,10 @@ func populateFuseArgs(opts *C.fuse_options_t, args *C.fuse_args_t) (*C.fuse_opti
 
 	if opts.allow_other {
 		options += ",allow_other"
+	}
+
+	if opts.non_empty {
+		options += ",nonempty"
 	}
 
 	if opts.readonly {