+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 ## 2.0.0-preview.5 (WIP)
 **Bug Fixes**
-- 
+- [#968](https://github.com/Azure/azure-storage-fuse/issues/968) Duplicate directory listing
+- [#964](https://github.com/Azure/azure-storage-fuse/issues/964) Rename for FNS account failing with source does not exists error
+- [#972](https://github.com/Azure/azure-storage-fuse/issues/972) Mount all fails 
+- Added support for "-o nonempty" to mount on a non-empty mount path
+
+**Breaking Changes**
+- Defaults for retry policy changed. Max retries: 3 to 5, Retry delay: 3600 to 900 seconds, Max retry delay: 4 to 60 seconds
+
+**Features**
+- Added new CLI parameter "--subdirectory=" to mount only a subdirectory from given container 
+
 
 ## 2.0.0-preview.4 (2022-11-03)
 **Breaking Changes**
@@ -10,6 +20,8 @@
 **Features**
 - Added support for health-monitor stop --pid=<pid> and health-monitor stop all commands
 - Added support to work with virtual directories without special marker directory using the virtual-directory config option.
+- Added support for object ID support for MSI credentials
+- Added support for system assigned IDs for MSI credentials
 
 **Bug Fixes**
 - Auto detect auth mode based on storage config